Corrections from code review. Splitting amazon relations logic into separate file and adding a few comments
